// super/lib/tpstreams/types.ts
export interface TpPaginatedResponse<T> {
  readonly count: number;
  readonly next: string | null;
  readonly previous: string | null;
  readonly results: ReadonlyArray<T>;
}
// ...
export interface TpUsageRecord {
  readonly date: string;
  readonly bandwidth_used_bytes: number;
  readonly active_storage_bytes: number;
  readonly deleted_storage_bytes: number;
  readonly total_storage_bytes: number;
  readonly subtitle_generation_cost: number;
  readonly live_stream_usage_seconds?: number;
}